django-backend/admana (copy)/requirementst.txt
django-url-shortening
